Refactor Article and Page with concerns
