+++ app/models/article.rb
@@ -1,6 +1,6 @@
 class Article < ApplicationRecord
-  belongs_to :user, optional: true
-  belongs_to :status
+  include Post
+
   belongs_to :theme
 
   has_many :taggings, dependent: :destroy
@@ -11,28 +11,12 @@ class Article < ApplicationRecord
 
   accepts_nested_attributes_for :contributions, reject_if: :all_blank, allow_destroy: true
 
-  scope :draft,       -> { where(status: Status.find_by(name: "draft")) }
-  scope :edited,      -> { where(status: Status.find_by(name: "edited")) }
-  scope :designed,    -> { where(status: Status.find_by(name: "designed")) }
-  scope :published,   -> { where(status: Status.find_by(name: "published")) }
-
-  scope :live,        -> { where("published_at < ?", Time.now) }
-
-  before_validation :generate_slug,            on: [:create, :update]
   before_validation :generate_published_dates, on: [:create, :update]
-  before_validation :generate_draft_code,      on: [:create, :update]
   before_validation :downcase_content_format,  on: [:create, :update]
 
   default_scope { order("published_at DESC") }
+  scope :live,        -> { where("published_at < ?", Time.now) }
   scope :on, lambda { |date| where("published_at BETWEEN ? AND ?", date.try(:beginning_of_day), date.try(:end_of_day)) }
-
-  def name
-    if title.present? && subtitle.present?
-      "#{title} : #{subtitle}"
-    else
-      title
-    end
-  end
 
   def path
     if published?
@@ -42,6 +26,8 @@ class Article < ApplicationRecord
     end
   end
 
+ # Overwrites slug_exists? from Slug.  We allow duplicate slugs on different
+ # published_at dates.
   def slug_exists?
     Article.on(published_at).where(slug: slug).exists?
   end
@@ -68,62 +54,7 @@ class Article < ApplicationRecord
     end
   end
 
-  # article states through the process from creation to publishing
-  def draft?
-    status.name == "draft"
-  end
-
-  def edited?
-    status.name == "edited"
-  end
-
-  def designed?
-    status.name == "designed"
-  end
-
-  def published?
-    status.name == "published"
-  end
-
-  def dated?
-    published_at.present?
-  end
-
-  def meta_description
-    if summary.blank?
-      html = Kramdown::Document.new(
-        content,
-        input: :kramdown,
-        remove_block_html_tags: false,
-        transliterated_header_ids: true
-      ).to_html.to_s
-
-      doc = Nokogiri::HTML(html)
-      doc.css("body").text.truncate(200)
-    else
-      summary
-    end
-  end
-
   private
-
-  def generate_slug
-    if self.new_record? || self.slug_changed? || self.slug.blank?
-      n = 0
-
-      if slug.blank?
-        self.slug = self.name.to_slug
-      end
-
-      while slug_exists?
-        self.slug = name
-        n += 1
-        "#{self.slug} #{n}".to_slug
-      end
-    end
-
-    self.slug = self.slug.to_slug
-  end
 
   def generate_published_dates
     if published_at.present?
@@ -131,10 +62,6 @@ class Article < ApplicationRecord
       self.month = published_at.month.to_s.rjust(2, "0") if published_at.month.present?
       self.day   = published_at.day.to_s.rjust(2, "0")   if published_at.day.present?
     end
-  end
-
-  def generate_draft_code
-    self.draft_code ||= SecureRandom.hex
   end
 
   def downcase_content_format
