+++ app/models/category.rb
@@ -1,15 +1,12 @@
 class Category < ApplicationRecord
+  include Slug
+
   has_many :categorizations, dependent: :destroy
   has_many :articles, through: :categorizations
 
   validates :name, presence: true, uniqueness: true
 
-  before_validation :generate_slug, on: [:create, :update]
   before_validation :strip_whitespace, on: [:create, :update]
-
-  def generate_slug
-    self.slug = name.to_slug
-  end
 
   def strip_whitespace
     self.name = name.strip
