+++ app/models/concerns/name_from_title.rb
@@ -1,0 +1,11 @@
+module NameFromTitle
+  extend ActiveSupport::Concern
+
+  def name
+    if title.present? && subtitle.present?
+      "#{title} : #{subtitle}"
+    else
+      title
+    end
+  end
+end
