+++ app/models/setting.rb
@@ -1,6 +1,7 @@
 class Setting < ApplicationRecord
+  include Slug
+
   validates :name, presence: true, uniqueness: true
-  before_validation :generate_slug, on: [:create, :update]
 
   def content
     saved_content.blank? ? fallback : saved_content
@@ -9,6 +10,6 @@ class Setting < ApplicationRecord
   private
 
   def generate_slug
-    self.slug = self.name.to_slug.gsub("-", "_")
+    self.slug = super.gsub("-", "_")
   end
 end
