+++ app/models/video.rb
@@ -1,14 +1,8 @@
 class Video < ApplicationRecord
+  include NameFromTitle
+
   default_scope { order("published_at DESC") }
 
-  def name
-    if title.present? && subtitle.present?
-      "#{title} : #{subtitle}"
-    else
-      title
-    end
-  end
-      
   def path
     "/videos/#{slug}"
   end
